Break out database manager constructors into submodule.
