+++ indra_db/util/__init__.py
@@ -6,11 +6,9 @@ Some key functions' capabilities include:
 - distilling and deleting statements
 """
 
-__all__ = ['get_defaults', 'get_primary_db', 'get_db', 'insert_agents',
-           'insert_pa_stmts', 'insert_db_stmts', 'get_raw_stmts_frm_db_list',
-           'distill_stmts', 'regularize_agent_id']
+__all__ = ['get_primary_db', 'get_db', 'insert_agents', 'insert_pa_stmts',
+           'insert_db_stmts', 'get_raw_stmts_frm_db_list', 'distill_stmts', 'regularize_agent_id']
 
-import re
 import json
 import zlib
 import pickle
@@ -25,108 +23,16 @@ from indra.util.nested_dict import NestedDict
 from indra.util.get_version import get_version
 from indra.util import batch_iter, clockit
 from indra.statements import Complex, SelfModification, ActiveForm,\
-    stmts_from_json, Conversion, Translocation, Statement
+    Conversion, Translocation, Statement
 
-from indra_db.managers.database_manager import DatabaseManager
 from indra_db.exceptions import IndraDbException
-from indra_db.config import get_databases as get_defaults
-
-logger = logging.getLogger('db_util')
 
 
-__PRIMARY_DB = None
+from .dump_sif import *
+from .constructors import *
+from .content_scripts import *
 
-
-def get_primary_db(force_new=False):
-    """Get a DatabaseManager instance for the primary database host.
-
-    The primary database host is defined in the defaults.txt file, or in a file
-    given by the environment variable DEFAULTS_FILE. Alternatively, it may be
-    defined by the INDRADBPRIMARY environment variable. If none of the above
-    are specified, this function will raise an exception.
-
-    Note: by default, calling this function twice will return the same
-    `DatabaseManager` instance. In other words::
-
-        db1 = get_primary_db()
-        db2 = get_primary_db()
-        db1 is db2
-
-    This means also that, for example `db1.select_one(db2.TextRef)` will work,
-    in the above context.
-
-    It is still recommended that when creating a script or function, or other
-    general application, you should not rely on this feature to get your access
-    to the database, as it can make substituting a different database host both
-    complicated and messy. Rather, a database instance should be explicitly
-    passed between different users as is done in `get_statements_by_gene_role_type`
-    function's call to `get_statements` in `indra.db.query_db_stmts`.
-
-    Parameters
-    ----------
-    force_new : bool
-        If true, a new instance will be created and returned, regardless of
-        whether there is an existing instance or not. Default is False, so that
-        if this function has been called before within the global scope, a the
-        instance that was first created will be returned.
-
-    Returns
-    -------
-    primary_db : :py:class:`DatabaseManager`
-        An instance of the database manager that is attached to the primary
-        database.
-    """
-    defaults = get_defaults()
-    if 'primary' in defaults.keys():
-        primary_host = defaults['primary']
-    else:
-        raise IndraDbException("No primary host available in defaults file.")
-
-    global __PRIMARY_DB
-    if __PRIMARY_DB is None or force_new:
-        __PRIMARY_DB = DatabaseManager(primary_host, label='primary')
-        __PRIMARY_DB.grab_session()
-    return __PRIMARY_DB
-
-
-def get_test_db():
-    """Get a DatabaseManager for the test database."""
-    defaults = get_defaults()
-    test_defaults = {k: v for k, v in defaults.items() if 'test' in k}
-    key_list = list(test_defaults.keys())
-    key_list.sort()
-    db = None
-    for k in key_list:
-        test_name = test_defaults[k]
-        m = re.match('(\w+)://.*?/([\w.]+)', test_name)
-        if m is None:
-            logger.warning("Poorly formed db name: %s" % test_name)
-            continue
-        sqltype = m.groups()[0]
-        try:
-            db = DatabaseManager(test_name, sqltype=sqltype, label=k)
-            db.grab_session()
-        except Exception as e:
-            logger.error("%s didn't work" % test_name)
-            logger.exception(e)
-            continue  # Clearly this test database won't work.
-        logger.info("Using test database %s." % k)
-        break
-    if db is None:
-        logger.error("Could not find any test database names.")
-    return db
-
-
-def get_db(db_label):
-    """Get a db instance base on it's name in the config or env."""
-    defaults = get_defaults()
-    db_name = defaults[db_label]
-    m = re.match('(\w+)://.*?/([\w.]+)', db_name)
-    if m is None:
-        logger.error("Poorly formed db name: %s" % db_name)
-        return
-    sqltype = m.groups()[0]
-    return DatabaseManager(db_name, sqltype=sqltype, label=db_label)
+logger = logging.getLogger('util')
 
 
 @clockit
